+++ api/src/features/todo/todo_feature.py
@@ -4,16 +4,26 @@ from starlette.responses import JSONResponse
 from authentication.authentication import auth_with_jwt
 from authentication.models import User
 from common.responses import create_response
-from data_providers.get_repository import get_todo_repository
-from data_providers.repository_interfaces.todo_repository_interface import (
-    TodoRepositoryInterface,
+from features.todo.repository.todo_repository import get_todo_repository
+from features.todo.repository.todo_repository_interface import TodoRepositoryInterface
+from features.todo.use_cases.add_todo import (
+    AddTodoRequest,
+    AddTodoResponse,
+    add_todo_use_case,
 )
-
-from .use_cases.add_todo import AddTodoRequest, AddTodoResponse, add_todo_use_case
-from .use_cases.delete_todo_by_id import DeleteTodoByIdResponse, delete_todo_use_case
-from .use_cases.get_todo_all import GetTodoAllResponse, get_todo_all_use_case
-from .use_cases.get_todo_by_id import GetTodoByIdResponse, get_todo_by_id_use_case
-from .use_cases.update_todo import (
+from features.todo.use_cases.delete_todo_by_id import (
+    DeleteTodoByIdResponse,
+    delete_todo_use_case,
+)
+from features.todo.use_cases.get_todo_all import (
+    GetTodoAllResponse,
+    get_todo_all_use_case,
+)
+from features.todo.use_cases.get_todo_by_id import (
+    GetTodoByIdResponse,
+    get_todo_by_id_use_case,
+)
+from features.todo.use_cases.update_todo import (
     UpdateTodoRequest,
     UpdateTodoResponse,
     update_todo_use_case,
@@ -25,9 +35,9 @@ router = APIRouter(tags=["todos"], prefix="/todos")
 @router.post("", operation_id="create")
 @create_response(JSONResponse)
 def add_todo(
-    data: AddTodoRequest,
-    user: User = Depends(auth_with_jwt),
-    todo_repository: TodoRepositoryInterface = Depends(get_todo_repository),
+        data: AddTodoRequest,
+        user: User = Depends(auth_with_jwt),
+        todo_repository: TodoRepositoryInterface = Depends(get_todo_repository),
 ) -> AddTodoResponse:
     return add_todo_use_case(data=data, user_id=user.user_id, todo_repository=todo_repository)
 
@@ -35,9 +45,9 @@ def add_todo(
 @router.get("/{id}", operation_id="get_by_id")
 @create_response(JSONResponse)
 def get_todo_by_id(
-    id: str,
-    user: User = Depends(auth_with_jwt),
-    todo_repository: TodoRepositoryInterface = Depends(get_todo_repository),
+        id: str,
+        user: User = Depends(auth_with_jwt),
+        todo_repository: TodoRepositoryInterface = Depends(get_todo_repository),
 ) -> GetTodoByIdResponse:
     return get_todo_by_id_use_case(id=id, user_id=user.user_id, todo_repository=todo_repository)
 
@@ -45,9 +55,9 @@ def get_todo_by_id(
 @router.delete("/{id}", operation_id="delete_by_id")
 @create_response(JSONResponse)
 def delete_todo_by_id(
-    id: str,
-    user: User = Depends(auth_with_jwt),
-    todo_repository: TodoRepositoryInterface = Depends(get_todo_repository),
+        id: str,
+        user: User = Depends(auth_with_jwt),
+        todo_repository: TodoRepositoryInterface = Depends(get_todo_repository),
 ) -> DeleteTodoByIdResponse:
     return delete_todo_use_case(id=id, user_id=user.user_id, todo_repository=todo_repository)
 
@@ -55,7 +65,7 @@ def delete_todo_by_id(
 @router.get("", operation_id="get_all")
 @create_response(JSONResponse)
 def get_todo_all(
-    user: User = Depends(auth_with_jwt), todo_repository: TodoRepositoryInterface = Depends(get_todo_repository)
+        user: User = Depends(auth_with_jwt), todo_repository: TodoRepositoryInterface = Depends(get_todo_repository)
 ) -> list[GetTodoAllResponse]:
     return get_todo_all_use_case(user_id=user.user_id, todo_repository=todo_repository)
 
@@ -63,9 +73,9 @@ def get_todo_all(
 @router.put("/{id}", operation_id="update_by_id")
 @create_response(JSONResponse)
 def update_todo(
-    id: str,
-    data: UpdateTodoRequest,
-    user: User = Depends(auth_with_jwt),
-    todo_repository: TodoRepositoryInterface = Depends(get_todo_repository),
+        id: str,
+        data: UpdateTodoRequest,
+        user: User = Depends(auth_with_jwt),
+        todo_repository: TodoRepositoryInterface = Depends(get_todo_repository),
 ) -> UpdateTodoResponse:
     return update_todo_use_case(id=id, data=data, user_id=user.user_id, todo_repository=todo_repository)
