+++ api/src/features/todo/use_cases/add_todo.py
@@ -2,10 +2,8 @@ import uuid
 
 from pydantic import BaseModel, Field
 
-from data_providers.repository_interfaces.todo_repository_interface import (
-    TodoRepositoryInterface,
-)
-from entities.todo_item import TodoItem
+from features.todo.entities.todo_item import TodoItem
+from features.todo.repository.todo_repository_interface import TodoRepositoryInterface
 
 
 class AddTodoRequest(BaseModel):
