+++ api/src/features/todo/use_cases/get_todo_by_id.py
@@ -3,10 +3,8 @@ from typing import cast
 from pydantic import BaseModel
 
 from common.exceptions import MissingPrivilegeException
-from data_providers.repository_interfaces.todo_repository_interface import (
-    TodoRepositoryInterface,
-)
-from entities.todo_item import TodoItem
+from features.todo.entities.todo_item import TodoItem
+from features.todo.repository.todo_repository_interface import TodoRepositoryInterface
 
 
 class GetTodoByIdResponse(BaseModel):
