+++ api/src/features/todo/use_cases/update_todo.py
@@ -1,10 +1,8 @@
 from pydantic import BaseModel, Field
 
 from common.exceptions import MissingPrivilegeException
-from data_providers.repository_interfaces.todo_repository_interface import (
-    TodoRepositoryInterface,
-)
-from entities.todo_item import TodoItem
+from features.todo.entities.todo_item import TodoItem
+from features.todo.repository.todo_repository_interface import TodoRepositoryInterface
 
 
 class UpdateTodoRequest(BaseModel):
